removed println, moved createJWT to a separate function
